test: 컴포넌트 스토리 업데이트
