+++ .storybook/main.ts
@@ -16,4 +16,5 @@ const config: StorybookConfig = {
   },
   staticDirs: ["..\\public"],
 };
+
 export default config;
