+++ src/component/CardDetailPage/CardDetailForm/CardNumberInput/CardNumberInput.stories.tsx
@@ -1,20 +1,21 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import CardNumberInput from "./CardNumberInput";
+import { CardDetailProvider } from "../../../../context/CardDetailContext";
 
 const meta = {
   component: CardNumberInput,
   title: "CardDetailForm/CardNumberInput",
+  decorators: [
+    (Story) => (
+      <CardDetailProvider>
+        <Story />
+      </CardDetailProvider>
+    ),
+  ],
 } satisfies Meta<typeof CardNumberInput>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary: Story = {
-  args: {
-    changeCardNumber: (e: React.ChangeEvent<HTMLInputElement>) => {
-      console.log(e.target.value);
-    },
-    cardNumberHidden: "1111-2222-••••-••••",
-  },
-};
+export const Primary: Story = {};
