+++ src/component/CardDetailPage/CardDetailForm/CardPasswordInput/CardPasswordInput.stories.tsx
@@ -1,20 +1,21 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import CardPasswordInput from "./CardPasswordInput";
+import { CardDetailProvider } from "../../../../context/CardDetailContext";
 
 const meta = {
   component: CardPasswordInput,
   title: "CardDetailForm/CardPasswordInput",
+  decorators: [
+    (Story) => (
+      <CardDetailProvider>
+        <Story />
+      </CardDetailProvider>
+    ),
+  ],
 } satisfies Meta<typeof CardPasswordInput>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary: Story = {
-  args: {
-    changeCardPassword: (e: React.FormEvent<HTMLInputElement>) => {
-      console.log(e.currentTarget.value);
-    },
-    cardPassword: ["1", "2"],
-  },
-};
+export const Primary: Story = {};
